# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PartSignal API
  version: 0.1.1
  description: PartSignal MVP 契约。除登录外，写请求必须携带 X-CSRF-Token。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/content-versions/{content_version_id}/publication-package:
    get:
      tags: [publication]
      operationId: getPublicationPackage
      parameters: [{$ref: '#/components/parameters/ContentVersionId'}]
      responses:
        '200':
          description: 由批准 Markdown 派生的人工发布包
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicationPackage'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
...
  /api/v1/platform-accounts:
    get:
      tags: [publication]
      operationId: listPlatformAccounts
      responses:
        '200':
          description: 平台账号标识
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlatformAccountList'}
    post:
      tags: [publication]
      operationId: createPlatformAccount
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PlatformAccountCreate'}
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlatformAccount'}
  /api/v1/publication-records/manual:
    post:
      tags: [publication]
      operationId: createManualPublication
      parameters:
        - {$ref: '#/components/parameters/CsrfHeader'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ManualPublicationCreate'}
      responses:
        '201':
          description: 已创建人工发布记录
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicationRecord'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/publication-records:
    get:
      tags: [publication]
      operationId: listPublicationRecords
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/PublicationStatus'}}
      responses:
        '200':
          description: 发布记录列表
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicationRecordList'}
  /api/v1/publication-records/{publication_id}:
    get:
      tags: [publication]
      operationId: getPublicationRecord
      parameters: [{$ref: '#/components/parameters/PublicationId'}]
      responses:
        '200':
          description: 发布记录
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicationRecord'}
  /api/v1/publication-records/{publication_id}/{command}:
    post:
      tags: [publication]
      operationId: commandPublicationRecord
      parameters:
        - {$ref: '#/components/parameters/PublicationId'}
        - name: command
          in: path
          required: true
          schema:
            type: string
            enum: [mark-platform-review, mark-published, verify, reject, remove, mark-verification-failed]
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PublicationCommand'}
      responses:
        '200':
          description: 已执行状态命令
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PublicationRecord'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: partsignal_session
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 32}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 128}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    ContentVersionId: {name: content_version_id, in: path, required: true, schema: {type: string, format: uuid}}
    PublicationId: {name: publication_id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    ErrorResponse:
      description: 业务或校验错误
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorDetail'}
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true, default: {}}
        request_id: {type: string}
...
    Confidentiality: {type: string, enum: [PUBLIC, INTERNAL, RESTRICTED]}
...
    PublicationPackage:
      type: object
      additionalProperties: false
      required: [content_version_id, fact_version_id, title, body_markdown, body_html, body_text, tags, canonical_url, content_hash]
      properties:
        content_version_id: {type: string, format: uuid}
        fact_version_id: {type: string, format: uuid}
        title: {type: string}
        body_markdown: {type: string}
        body_html: {type: string}
        body_text: {type: string}
        tags: {type: array, items: {type: string}}
        canonical_url: {type: string, format: uri, minLength: 1, maxLength: 2083}
        content_hash: {type: string}
    PlatformAccountCreate:
      type: object
      additionalProperties: false
      required: [platform_profile_id, label, account_identifier]
      properties:
        platform_profile_id: {type: string, format: uuid}
        label: {type: string}
        account_identifier: {type: string}
    PlatformAccount:
      allOf:
        - {$ref: '#/components/schemas/PlatformAccountCreate'}
        - type: object
          required: [id, is_active]
          properties:
            id: {type: string, format: uuid}
            is_active: {type: boolean}
    PlatformAccountList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/PlatformAccount'}}
    ManualPublicationCreate:
      type: object
      additionalProperties: false
      required: [content_version_id, platform_account_id, section_url]
      properties:
        content_version_id: {type: string, format: uuid}
        platform_account_id: {type: string, format: uuid}
        section_url: {type: string, format: uri, minLength: 1, maxLength: 2083}
        attachment_file_ids: {type: array, uniqueItems: true, items: {type: string, format: uuid}}
    PublicationStatus:
      type: string
      enum: [PENDING_MANUAL_PUBLISH, PLATFORM_REVIEW, PUBLISHED, VERIFIED, REJECTED, REMOVED, VERIFICATION_FAILED]
    PublicationCommand:
      type: object
      additionalProperties: false
      required: [comment]
      properties:
        actual_title: {type: [string, 'null']}
        final_url: {type: [string, 'null'], format: uri, minLength: 1, maxLength: 2083}
        published_at: {type: [string, 'null'], format: date-time}
        content_matches: {type: [boolean, 'null']}
        comment: {type: string}
    PublicationRecord:
      type: object
      additionalProperties: false
      required: [id, content_version_id, platform_account_id, section_url, status, content_hash, created_by, created_at, status_events, attachments]
      properties:
        id: {type: string, format: uuid}
        content_version_id: {type: string, format: uuid}
        platform_account_id: {type: string, format: uuid}
        section_url: {type: string, format: uri, minLength: 1, maxLength: 2083}
        actual_title: {type: [string, 'null']}
        final_url: {type: [string, 'null'], format: uri, minLength: 1, maxLength: 2083}
        published_at: {type: [string, 'null'], format: date-time}
        status: {$ref: '#/components/schemas/PublicationStatus'}
        content_hash: {type: string}
        created_by: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
        status_events: {type: array, items: {$ref: '#/components/schemas/PublicationEvent'}}
        attachments: {type: array, items: {$ref: '#/components/schemas/FileRecord'}}
    PublicationRecordList:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/PublicationRecord'}}
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}
    PublicationEvent:
      type: object
      additionalProperties: false
      required: [status, comment, actor_id, created_at]
      properties:
        status: {$ref: '#/components/schemas/PublicationStatus'}
        comment: {type: string}
        actor_id: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
...
    FileCategory: {type: string, enum: [EVIDENCE, OPERATION_SCREENSHOT, PUBLICATION_ASSET]}
...
    FileRecord:
      type: object
      additionalProperties: false
      required: [id, category, original_filename, object_key, content_type, size, sha256, access_level, status, created_at]
      properties:
        id: {type: string, format: uuid}
        category: {$ref: '#/components/schemas/FileCategory'}
        original_filename: {type: string}
        object_key: {type: string}
        content_type: {type: string}
        size: {type: integer}
        sha256: {type: string}
        access_level: {$ref: '#/components/schemas/Confidentiality'}
        status: {type: string, enum: [PENDING, VERIFIED, FAILED, ABORTED]}
        created_at: {type: string, format: date-time}
        verified_at: {type: [string, 'null'], format: date-time}
...